+++ app/javascript/membership_verify.js
