+++ app/javascript/modules/ajax_error_notification.js.coffee
