+++ app/javascript/modules/ajax_replace.js.coffee
