+++ app/javascript/modules/ajax_upload.js.coffee
