+++ app/javascript/modules/async_downloads.js.coffee
