+++ app/javascript/modules/auto_submit.js.coffee
