+++ app/javascript/modules/clear_input.js.coffee
