+++ app/javascript/modules/clickable_placeholders.js.coffee
