+++ app/javascript/modules/collapse_overflow_fix.js.coffee
