+++ app/javascript/modules/copy_to_clipboard.js.coffee
