+++ app/javascript/modules/disabled_links.js.coffee
