+++ app/javascript/modules/element_swapper.js.coffee
