+++ app/javascript/modules/element_toggler.js.coffee
