+++ app/javascript/modules/event_kind_preconditions.js.coffee
