+++ app/javascript/modules/events/date_period_validator.js.coffee
