+++ app/javascript/modules/events/event_tags.js.coffee
