+++ app/javascript/modules/events/events.js.coffee
