+++ app/javascript/modules/fields_autofocus.js.coffee
