+++ app/javascript/modules/groups/group_contact_toggle.js.coffee
