+++ app/javascript/modules/input_enabler.js.coffee
