+++ app/javascript/modules/invitation_lists.js.coffee
