+++ app/javascript/modules/invoice_message_template.js
