+++ app/javascript/modules/invoices.js.coffee
