+++ app/javascript/modules/mailing_list_labels.js.coffee
