+++ app/javascript/modules/multiselect_add_chips.js.coffee
