+++ app/javascript/modules/notes.js.coffee
