+++ app/javascript/modules/participation_lists.js.coffee
