+++ app/javascript/modules/people/people_filter_attribute.js.coffee
