+++ app/javascript/modules/people/people_filter_role_toggle.js.coffee
