+++ app/javascript/modules/people/toggle_household_for_labels.js
