+++ app/javascript/modules/persistent_dropdown.js.coffee
