+++ app/javascript/modules/profile_overlay.coffee
