+++ app/javascript/modules/spinner.js.coffee
