+++ app/javascript/modules/string_trim.js.coffee
