+++ app/javascript/modules/subscriber_lists.js.coffee
