+++ app/views/people/membership/verify/show.html.haml
@@ -10,8 +10,8 @@
     %meta{charset: 'utf-8'}
     %title= "#{Settings.application.name} - #{title}"
     %meta{name: 'viewport', content: 'width=device-width, initial-scale=1.0'}
-    = stylesheet_pack_tag 'application'
-    = javascript_pack_tag 'membership_verify'
+    = stylesheet_link_tag 'application'
+    = javascript_include_tag 'membership_verify'
   %body
     #membership-verify
       = render partial: 'header'
